% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanCerebellumSubset.R
\name{HumanCerebellumNormSubset}
\alias{HumanCerebellumNormSubset}
\title{HumanCerebellumSubset}
\usage{
HumanCerebellumNormSubset()
}
\value{
A `SpatialExperiment` object containing the subsetted normalized
human cerebellum spatial transcriptomics data.
}
\description{
A subset of the Visium spatial 10X Genomics spatial gene
expression data for the human cerebellum, after quality control filtering
and normalization. The original data was downloaded using the function
[`HumanCerebellum`][TENxVisiumData::HumanCerebellum] from the Bioconductor
package [`TENxVisiumData`][TENxVisiumData::TENxVisiumData]. The script
`inst/scripts/prepare_HumanCerebellumSubset.R` was used to perform the
quality control filtering and normalization, and to save the resulting
subsetted data in different files in the `inst/extdata` directory.
}
\examples{

spe <- HumanCerebellumNormSubset()
spe

}
