\name{GSVAdata-package}
\alias{GSVAdata-package}
\alias{GSVAdata}
\docType{data}
\title{
Data employed in the vignette of the GSVA package.
}
\description{
This package contains data employed in the vignette of the GSVA package.
}

\section{Data sets}{
  \itemize{
    \item \code{\link{leukemia}} Leukemia data by Armstrong et al. (2002) from
          the Broad Institute.
    \item \code{\link{c2BroadSets}} C2 canonical pathways from the MSigDB 3.0
          database of gene sets at the Broad Institute.
  }
}

\author{
S. Haenzelmann, J. Guinney and R. Castelo
}

\references{
Scott A. Armstrong, Jane E. Staunton, Lewis B. Silverman, Rob Pieters, Monique L. den Boer, Mark D. Minden, Stephen E. Sallan, Eric S. Lander, Todd R. Golub and Stanley J. Korsmeyer. MLL translocations specify a distinct gene expression profile that distinguishes a unique leukemia. \emph{Nat Genet} 30:41-47, 2002.
}

\keyword{dataset}
